feat: refactor loan application logic and remove unused components
