--- src/types/account.ts
@@ -1,15 +1,0 @@
-export type LoanStatus = 'Pending' | 'Approved' | 'Declined' | 'Undecided';
-
-export interface AccountData {
-  email?: string;
-  taxId: string;
-  businessName: string;
-  requestedAmount: number;
-  decision: LoanStatus;
-}
-
-export interface LoanResponse {
-  details: { loanDecision: string }[];
-  message: string;
-  statusCode: number;
-}
